Add files via upload
The following code is used to generate point data for annual ET maps. Pixel data from Landsat are temporally interpolated as points which can then be spatially interpolated (e.g kriging). Depending on the region of interest, the for loop which is used to interpolate the data can take a long time to process. Sometimes, I've left it run overnight. The loop can be divided into sections if necessary.